requirements.txt
blinker==1.9.0
click==8.1.8
EbookLib==0.18
Flask==3.1.0
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.3.1
MarkupSafe==3.0.2
PyPDF2==3.0.1
python-docx==1.1.2
six==1.17.0
typing_extensions==4.12.2
Werkzeug==3.1.3
